Stabilize Playwright CI and Allure reporting
