+++ playwright-automation/tests/ui/smoke/data_connection_e2e.spec.js
@@ -13,23 +13,44 @@ test.beforeEach(async ({ masterFixture }) => {
   dataImport = masterFixture.dataImport;
   await appPage.navigateToUrl(editorUrl);
   await appPage.hardWait(5000);
-  console.log('Actions', dataImport);
 });
-test.describe('@smoke Media Licenses', () => {
-  test('Login as editor user', async ({ page }) => {
+
+test.describe.serial('@smoke Add and Delete Data Connection', () => {
+  test('Verify adding new data connection', async ({ page }) => {
     await page.goto(editorUrl);
     await dataImport.navigateToDataImport();
-    await dataImport.navigatetoDataLoctn();
-    await dataImport.clickAddNewDataLoctn();
+    await dataImport.navigateToDataCnctn();
+    await dataImport.clickAddNewDataCntn();
 
     const DataLName = `Automation Test Data`;
-    const DataLDescription = `Automation DEscription Data for testing `;
+    const DataLDescription = `Automation Description Data for testing `;
     const DatarandomNmb = Math.floor(Math.random() * 100) + 1;
     const DataSortOrder = `${DatarandomNmb}`;
 
     await dataImport.enterIngestDetails(DataLName, DataLDescription, DataSortOrder);
+    await dataImport.dropdownfield('NAS');
 
-    await dataImport.dropdownValue('Local Volume - Clips');
+    await expect(await dataImport.validateMessage()).toBe(true);
+
+    await dataImport.backbtndataC();
+    await appPage.logOut();
+  });
+
+  test('Verify Search and Delete for Data Connections', async ({ page }) => {
+    await page.goto(editorUrl);
+    await dataImport.navigateToDataImport();
+    await dataImport.navigateToDataCnctn();
+
+    const keyword = `Automation Test Data`;
+
+    await dataImport.searchboxValue(keyword);
+    await dataImport.clickonrowValue();
+    await dataImport.clickDelete();
+    await dataImport.removeBtn();
+
+    await dataImport.navigatetoConnection();
+    await dataImport.validateDeleteMessage();
+
     await appPage.logOut();
   });
 });
